Initial commit for 0.3.0
This is a complete rewrite of this project with a better design and no
backward compatibility (sadly). While not yet coded the design aim at
providing a full complete life-cycle of the installed packages (like
major upgrade and backups). Beside one of the main target of this
redesign is to help with way easier maintenance as most of the
"functional" code is written in rhai. The goal is to write less Rust in
the future and so being able to evolve faster.